+++ packages/search/src/types/FlagKeyword.ts
@@ -1,0 +1,12 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+export default interface FlagKeyword {
+    keyword: string;
+    
+}
