+++ packages/search/src/types/StoreProps.ts
